atomisp2: css2400 css2401: CSS2.1/ISP2400 B0, 2401 A0 and 2401 legacy A0 ci_master_20131215_0443
BZ: 159309

Merge CSS2.1 for ISP2400 B0, 2401 A0 and 2401 A0 Legacy hardware. This is based ci_master_20131215_0443 firmware tag.

+++ drivers/media/pci/atomisp2/css2401a0_legacy_v21/runtime/event/src/event.c
@@ -49,8 +49,8 @@
  * Refer to "sw_event_public.h" for details.
  */
 bool ia_css_event_encode(
-	uint32_t	*in,
-	uint32_t	nr,
+	uint8_t	*in,
+	uint8_t	nr,
 	uint32_t	*out)
 {
 	bool ret;
